+++ src/android/CleverTapPlugin.java
@@ -16,8 +16,17 @@ import android.os.Bundle;
 import android.util.Log;
 import android.location.Location;
 
+import androidx.annotation.NonNull;
+import com.clevertap.android.sdk.PushPermissionResponseListener;
+import com.clevertap.android.sdk.inapp.CTInAppNotification;
+import com.clevertap.android.sdk.inapp.CTLocalInApp;
 import com.clevertap.android.sdk.pushnotification.CTPushNotificationListener;
 import com.clevertap.android.sdk.pushnotification.amp.CTPushAmpListener;
+import com.clevertap.android.sdk.variables.CTVariableUtils;
+import com.clevertap.android.sdk.variables.Var;
+import com.clevertap.android.sdk.variables.callbacks.VariableCallback;
+import com.clevertap.android.sdk.variables.callbacks.VariablesChangedCallback;
+import java.util.Objects;
 import java.util.Set;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaWebView;
@@ -25,6 +34,7 @@ import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.PluginResult;
 
+import org.apache.cordova.PluginResult.Status;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -62,12 +72,14 @@ import com.clevertap.android.sdk.interfaces.NotificationHandler;
 
 public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAppNotificationListener, CTInboxListener,
         InboxMessageButtonListener, InAppNotificationButtonListener, DisplayUnitListener,
-        CTFeatureFlagsListener, CTProductConfigListener, CTPushNotificationListener, CTPushAmpListener, InboxMessageListener {
+        CTFeatureFlagsListener, CTProductConfigListener, CTPushNotificationListener, CTPushAmpListener, InboxMessageListener,
+        PushPermissionResponseListener {
 
     private static final String LOG_TAG = "CLEVERTAP_PLUGIN";
     private static String CLEVERTAP_API_ERROR;
     private static CleverTapAPI cleverTap;
     private boolean callbackDone = false;
+    public static Map<String, Object> variables = new HashMap<>();
 
 
     @Override
@@ -86,6 +98,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         cleverTap.setCTProductConfigListener(this);
         cleverTap.setCTPushNotificationListener(this);
         cleverTap.setCTPushAmpListener(this);
+        cleverTap.registerPushPermissionNotificationResponseListener(this);
         cleverTap.setLibrary("Cordova");
 
         try {
@@ -1372,14 +1385,6 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 }
             });
             return true;
-        } else if (action.equals("deleteInboxMessagesForIds")) {// NO-OP for Android
-
-            PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
-            _result.setKeepCallback(true);
-            callbackContext.sendPluginResult(_result);
-
-            return true;
-
         } else if (action.equals("markReadInboxMessageForId")) {
             final String messageId = (args.length() == 1 ? args.getString(0) : "");
 
@@ -1392,6 +1397,65 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 }
             });
             return true;
+        }  else if (action.equals("markReadInboxMessagesForIds")) {
+            JSONArray jsonArray = null;
+            if (args.length() == 1) {
+                jsonArray = args.getJSONArray(0);
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+            if (!haveError) {
+                final JSONArray finalJsonArray = jsonArray;
+                cordova.getThreadPool().execute(() -> {
+                    try {
+                        cleverTap.markReadInboxMessagesForIDs((ArrayList<String>) toStringList(finalJsonArray));
+
+                        PluginResult _result = new PluginResult(Status.NO_RESULT);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    } catch (Exception e) {
+                        PluginResult _result = new PluginResult(Status.ERROR, e.getLocalizedMessage());
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+                return true;
+            }
+        }  else if (action.equals("deleteInboxMessagesForIds")) {
+            JSONArray jsonArray = null;
+            if (args.length() == 1) {
+                jsonArray = args.getJSONArray(0);
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+            if (!haveError) {
+                final JSONArray finalJsonArray = jsonArray;
+                cordova.getThreadPool().execute(() -> {
+                    try {
+                        cleverTap.deleteInboxMessagesForIDs((ArrayList<String>) toStringList(finalJsonArray));
+
+                        PluginResult _result = new PluginResult(Status.NO_RESULT);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    } catch (Exception e) {
+                        PluginResult _result = new PluginResult(Status.ERROR, e.getLocalizedMessage());
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+                return true;
+            }
+        } else if (action.equals("dismissInbox")) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    cleverTap.dismissAppInbox();
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
         } else if (action.equals("pushInboxNotificationViewedEventForId")) {
             final String messageId = (args.length() == 1 ? args.getString(0) : "");
 
@@ -1672,12 +1736,275 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 }
             });
             return true;
+        } else if (action.equals("promptPushPrimer")) {
+            JSONObject pushPrimerJsonObject = null;
+            if (args.length() == 1) {
+                if (!args.isNull(0)) {
+                    try {
+                        pushPrimerJsonObject = processPushPrimerArgument(args.getJSONObject(0));
+                        if (pushPrimerJsonObject == null)
+                        {
+                            haveError = true;
+                            errorMsg = "Invalid parameters in push primer config";
+                        }
+                    } catch (Exception e) {
+                        haveError = true;
+                        errorMsg = e.getLocalizedMessage();
+                    }
+                } else {
+                    haveError = true;
+                    errorMsg = "object passed to promptPushPrimer can not be null!";
+                }
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+
+            if (!haveError) {
+                final JSONObject finalPushPrimerJsonObject = pushPrimerJsonObject;
+                cordova.getThreadPool().execute(new Runnable() {
+                    public void run() {
+                        cleverTap.promptPushPrimer(finalPushPrimerJsonObject);
+                        PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+                return true;
+            }
+        } else if (action.equals("promptForPushPermission")) {
+            boolean showFallbackSettings = false;
+
+            if (args.length() == 1) {
+                showFallbackSettings = args.getBoolean(0);
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+            if (!haveError) {
+                final boolean finalShowFallbackSettings = showFallbackSettings;
+                cordova.getThreadPool().execute(new Runnable() {
+                    public void run() {
+                        cleverTap.promptForPushPermission(finalShowFallbackSettings);
+                        PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+                return true;
+            }
+        } else if (action.equals("isPushPermissionGranted")) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    boolean value = cleverTap.isPushPermissionGranted();
+                    PluginResult _result = new PluginResult(PluginResult.Status.OK, value);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
+            return true;
+        } else if (action.equals("setLibrary")) {
+            String libName = args.getString(0);
+            int libVersion = args.getInt(1);
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    cleverTap.setCustomSdkVersion(libName,libVersion);
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
+            return true;
+        } else if (action.equals("syncVariables")) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    cleverTap.syncVariables();
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
+            return true;
+        } else if (action.equals("syncVariablesinProd")) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    Log.d(LOG_TAG, "syncVariablesinProd is no-op in Android");
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
+            return true;
+        } else if (action.equals("fetchVariables")) {
+            cordova.getThreadPool().execute(() -> {
+                cleverTap.fetchVariables(isSuccess -> {
+                    PluginResult _result = new PluginResult(PluginResult.Status.OK, isSuccess);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                });
+
+            });
+            return true;
+        } else if (action.equals("defineVariables")) {
+            Map<String,Object> variablesMap = null;
+            if (args.length() == 1) {
+                if (!args.isNull(0)) {
+                    try {
+                        variablesMap = toMap(args.getJSONObject(0));
+                    } catch (Exception e) {
+                        haveError = true;
+                        errorMsg = e.getLocalizedMessage();
+                    }
+                } else {
+                    haveError = true;
+                    errorMsg = "object passed to defineVariables can not be null!";
+                }
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+
+            if (!haveError) {
+                final Map<String,Object> finalVariablesJsonObject = variablesMap;
+                cordova.getThreadPool().execute(() -> {
+                    for (Map.Entry<String, Object> entry : finalVariablesJsonObject.entrySet()) {
+                        String key = entry.getKey();
+                        Object value = entry.getValue();
+                        variables.put(key, cleverTap.defineVariable(key, value));
+                    }
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                });
+                return true;
+            }
+        } else if (action.equals("getVariable")) {
+            String variableName = null;
+
+            if (args.length() == 1) {
+                variableName = args.getString(0);
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+
+            if (!haveError) {
+                final String finalVariableName = variableName;
+                cordova.getThreadPool().execute(() -> {
+                    try {
+                        Object value = getVariableValue(finalVariableName);
+                        PluginResult _result = getPluginResult(Status.OK, value);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    } catch (Exception e) {
+                        PluginResult _result = new PluginResult(Status.ERROR, e.getLocalizedMessage());
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+                return true;
+            }
+        } else if (action.equals("getVariables")) {
+            cordova.getThreadPool().execute(() -> {
+                JSONObject jsonVariables = getVariablesAsJson();
+                PluginResult _result = new PluginResult(PluginResult.Status.OK, jsonVariables);
+                _result.setKeepCallback(true);
+                callbackContext.sendPluginResult(_result);
+            });
+            return true;
+        } else if (action.equals("onVariablesChanged")) {
+            cordova.getThreadPool().execute(() -> {
+                cleverTap.addVariablesChangedCallback(new VariablesChangedCallback() {
+                    @Override
+                    public void variablesChanged() {
+                        JSONObject jsonVariables = getVariablesAsJson();
+                        PluginResult _result = new PluginResult(PluginResult.Status.OK,jsonVariables);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+            });
+            return true;
+        } else if (action.equals("onValueChanged")) {
+
+            String variableName = null;
+
+            if (args.length() == 1) {
+                variableName = args.getString(0);
+            } else {
+                haveError = true;
+                errorMsg = "Expected 1 argument";
+            }
+
+            if (!haveError) {
+                final String finalVariableName = variableName;
+                cordova.getThreadPool().execute(() -> {
+                    try {
+                        if (variables.containsKey(finalVariableName)) {
+                            Var<Object> variable = (Var<Object>) variables.get(finalVariableName);
+                            variable.addValueChangedCallback(new VariableCallback<Object>() {
+                                @Override
+                                public void onValueChanged(final Var<Object> variable) {
+                                    Object value = getVariableValue(finalVariableName);
+                                    PluginResult _result = getPluginResult(Status.OK, value);
+                                    _result.setKeepCallback(true);
+                                    callbackContext.sendPluginResult(_result);
+
+                                }
+                            });
+                        }
+                    } catch (Exception e) {
+                        PluginResult _result = new PluginResult(Status.ERROR, e.getLocalizedMessage());
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
+                    }
+                });
+                return true;
+            }
+
         }
 
         result = new PluginResult(PluginResult.Status.ERROR, errorMsg);
         result.setKeepCallback(true);
         callbackContext.sendPluginResult(result);
         return true;
+    }
+
+    @NonNull
+    private JSONObject getVariablesAsJson() {
+        JSONObject jsonVariables = new JSONObject();
+        for (String key : variables.keySet()) {
+            try {
+                jsonVariables.put(key,getVariableValue(key));
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return jsonVariables;
+    }
+
+    @NonNull
+    private PluginResult getPluginResult(final Status ok, final Object value) {
+        PluginResult _result;
+        if (value instanceof Boolean) {
+           _result  = new PluginResult(ok, (Boolean) value);
+        } else if (value instanceof Double) {
+            _result  = new PluginResult(ok, ((Double) value).floatValue());
+        } else if (value instanceof Float) {
+            _result  = new PluginResult(ok, (Float) value);
+        } else if (value instanceof Integer) {
+            _result  = new PluginResult(ok, (Integer) value);
+        } else if (value instanceof Long) {
+            _result  = new PluginResult(ok, ((Long) value).intValue());
+        } else if (value instanceof String) {
+            _result  = new PluginResult(ok, (String) value);
+        } else if (value instanceof JSONObject) {
+            _result  = new PluginResult(ok, (JSONObject) value);
+        } else {
+            _result  = new PluginResult(PluginResult.Status.ERROR, "unknown value type");
+        }
+
+        return _result;
     }
 
     //DisplayUnitListener
@@ -1745,6 +2072,16 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         });
     }
 
+    @Override
+    public void onPushPermissionResponse(final boolean accepted) {
+        final String json = "{'accepted':" + accepted + "}";
+        webView.getView().post(new Runnable() {
+            public void run() {
+                webView.loadUrl("javascript:cordova.fireDocumentEvent('onCleverTapPushPermissionResponseReceived'," + json + ");");
+            }
+        });
+    }
+
     // SyncListener
     public void profileDataUpdated(JSONObject updates) {
 
@@ -1788,13 +2125,34 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         });
     }
 
-    public void onInboxItemClicked(CTInboxMessage message){
-        if(message != null &&  message.getData() != null){
+    @Override
+    public void onShow(CTInAppNotification inAppNotification) {
+        if(inAppNotification != null &&  inAppNotification.getJsonDescription() != null){
             //Read the values
-            final String json = "{'customExtras':" + message.getData().toString() + "}";
+            final String json = "{'customExtras':" + inAppNotification.getJsonDescription().toString() + "}";
             webView.getView().post(new Runnable() {
                 public void run() {
-                    webView.loadUrl("javascript:cordova.fireDocumentEvent('onCleverTapInboxItemClick'," + json + ");");
+                    webView.loadUrl("javascript:cordova.fireDocumentEvent('onCleverTapInAppNotificationShow'," + json + ");");
+                }
+            });
+        }
+    }
+
+    public void onInboxItemClicked(CTInboxMessage message, int contentPageIndex, int buttonIndex){
+        if(message != null &&  message.getData() != null){
+            //Read the values
+            JSONObject jsonObject = new JSONObject();
+            try {
+                jsonObject.put("data",message.getData());
+                jsonObject.put("contentPageIndex",contentPageIndex);
+                jsonObject.put("buttonIndex",buttonIndex);
+            } catch (JSONException e) {
+               Log.e(LOG_TAG,"Failed to parse inbox message.");
+            }
+
+            webView.getView().post(new Runnable() {
+                public void run() {
+                    webView.loadUrl("javascript:cordova.fireDocumentEvent('onCleverTapInboxItemClick'," + jsonObject + ");");
                 }
             });
         }
@@ -2059,16 +2417,20 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         return map;
     }
 
-    private JSONObject getJsonFromMap(Map<String, ?> map) throws JSONException {
-        JSONObject jsonData = new JSONObject();
-        for (String key : map.keySet()) {
-            Object value = map.get(key);
-            if (value instanceof Map<?, ?>) {
-                value = getJsonFromMap((Map<String, ?>) value);
+    private Object getVariableValue(String name) {
+        if (variables.containsKey(name)) {
+            Var<?> variable = (Var<?>) variables.get(name);
+            Object variableValue = variable.value();
+            Object value;
+            if (CTVariableUtils.DICTIONARY.equals(variable.kind())) {
+                value = new JSONObject((Map<String, Object>) variableValue);
+            } else {
+                value = variableValue;
             }
-            jsonData.put(key, value);
+            return value;
         }
-        return jsonData;
+        throw new IllegalArgumentException(
+                "Variable name = " + name + " does not exist. Make sure you set variable first.");
     }
 
     private JSONArray displayUnitListToJSONArray(ArrayList<CleverTapDisplayUnit> displayUnits) throws JSONException {
@@ -2134,5 +2496,150 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             }
         }
         return json;
+    }
+
+    private JSONObject processPushPrimerArgument(JSONObject jsonObject) {
+        CTLocalInApp.InAppType inAppType = null;
+        String titleText = null, messageText = null, positiveBtnText = null, negativeBtnText = null,
+                backgroundColor = null, btnBorderColor = null, titleTextColor = null, messageTextColor = null,
+                btnTextColor = null, imageUrl = null, btnBackgroundColor = null, btnBorderRadius = null;
+        boolean fallbackToSettings = false, followDeviceOrientation = false;
+
+        final Iterator<String> iterator = jsonObject.keys();
+        while (iterator.hasNext()) {
+            try {
+                final String configKey = iterator.next();
+                switch (configKey) {
+                    case "inAppType":
+                        inAppType = inAppTypeFromString(jsonObject.getString(configKey));
+                        break;
+                    case "titleText":
+                        titleText = jsonObject.getString(configKey);
+                        break;
+                    case "messageText":
+                        messageText = jsonObject.getString(configKey);
+                        break;
+                    case "followDeviceOrientation":
+                        followDeviceOrientation = jsonObject.getBoolean(configKey);
+                        break;
+                    case "positiveBtnText":
+                        positiveBtnText = jsonObject.getString(configKey);
+                        break;
+                    case "negativeBtnText":
+                        negativeBtnText = jsonObject.getString(configKey);
+                        break;
+                    case "fallbackToSettings":
+                        fallbackToSettings = jsonObject.getBoolean(configKey);
+                        break;
+                    case "backgroundColor":
+                        backgroundColor = jsonObject.getString(configKey);
+                        break;
+                    case "btnBorderColor":
+                        btnBorderColor = jsonObject.getString(configKey);
+                        break;
+                    case "titleTextColor":
+                        titleTextColor = jsonObject.getString(configKey);
+                        break;
+                    case "messageTextColor":
+                        messageTextColor = jsonObject.getString(configKey);
+                        break;
+                    case "btnTextColor":
+                        btnTextColor = jsonObject.getString(configKey);
+                        break;
+                    case "imageUrl":
+                        imageUrl = jsonObject.getString(configKey);
+                        break;
+                    case "btnBackgroundColor":
+                        btnBackgroundColor = jsonObject.getString(configKey);
+                        break;
+                    case "btnBorderRadius":
+                        btnBorderRadius = jsonObject.getString(configKey);
+                        break;
+                }
+            } catch (Throwable t) {
+                Log.e(LOG_TAG, "invalid parameters in push primer config" + t.getLocalizedMessage());
+                return null;
+            }
+        }
+
+        //creates the builder instance of localInApp with all the required parameters
+        CTLocalInApp.Builder.Builder6 builderWithRequiredParams = getLocalInAppBuilderWithRequiredParam(
+                inAppType, titleText, messageText, followDeviceOrientation, positiveBtnText, negativeBtnText
+        );
+
+        //adds the optional parameters to the builder instance
+        if (backgroundColor != null) {
+            builderWithRequiredParams.setBackgroundColor(backgroundColor);
+        }
+        if (btnBorderColor != null) {
+            builderWithRequiredParams.setBtnBorderColor(btnBorderColor);
+        }
+        if (titleTextColor != null) {
+            builderWithRequiredParams.setTitleTextColor(titleTextColor);
+        }
+        if (messageTextColor != null) {
+            builderWithRequiredParams.setMessageTextColor(messageTextColor);
+        }
+        if (btnTextColor != null) {
+            builderWithRequiredParams.setBtnTextColor(btnTextColor);
+        }
+        if (imageUrl != null) {
+            builderWithRequiredParams.setImageUrl(imageUrl);
+        }
+        if (btnBackgroundColor != null) {
+            builderWithRequiredParams.setBtnBackgroundColor(btnBackgroundColor);
+        }
+        if (btnBorderRadius != null) {
+            builderWithRequiredParams.setBtnBorderRadius(btnBorderRadius);
+        }
+        builderWithRequiredParams.setFallbackToSettings(fallbackToSettings);
+
+        JSONObject localInAppConfig = builderWithRequiredParams.build();
+        Log.i(LOG_TAG, "LocalInAppConfig for push primer prompt: " + localInAppConfig);
+        return localInAppConfig;
+
+
+    }
+
+    /**
+     * Creates an instance of the {@link CTLocalInApp.Builder.Builder6} with the required parameters.
+     *
+     * @return the {@link CTLocalInApp.Builder.Builder6} instance
+     */
+    private CTLocalInApp.Builder.Builder6 getLocalInAppBuilderWithRequiredParam(CTLocalInApp.InAppType inAppType,
+            String titleText,
+            String messageText,
+            boolean followDeviceOrientation,
+            String positiveBtnText,
+            String negativeBtnText) {
+        //throws exception if any of the required parameter is missing
+        if (inAppType == null || titleText == null || messageText == null || positiveBtnText == null
+                || negativeBtnText == null) {
+            throw new IllegalArgumentException("mandatory parameters are missing in push primer config");
+        }
+
+        CTLocalInApp.Builder builder = CTLocalInApp.builder();
+
+        return builder.setInAppType(inAppType)
+                .setTitleText(titleText)
+                .setMessageText(messageText)
+                .followDeviceOrientation(followDeviceOrientation)
+                .setPositiveBtnText(positiveBtnText)
+                .setNegativeBtnText(negativeBtnText);
+    }
+
+    //returns InAppType type from the given string
+    private CTLocalInApp.InAppType inAppTypeFromString(String inAppType) {
+        if (inAppType == null) {
+            return null;
+        }
+        switch (inAppType) {
+            case "half-interstitial":
+                return CTLocalInApp.InAppType.HALF_INTERSTITIAL;
+            case "alert":
+                return CTLocalInApp.InAppType.ALERT;
+            default:
+                return null;
+        }
     }
 }