+++ www/CleverTap.js
@@ -567,6 +567,22 @@ CleverTap.prototype.reset = function(){
     cordova.exec(null, null, "CleverTapPlugin", "reset", []);
 }
 
+/****************************
+ * Android 13 Push Primer
+ ****************************/
+ 
+CleverTap.prototype.promptPushPrimer = function(localInAppObject){
+    cordova.exec(null, null, "CleverTapPlugin", "promptPushPrimer", [localInAppObject]);
+}
+
+CleverTap.prototype.promptForPushPermission = function(showFallbackSettings){
+    cordova.exec(null, null, "CleverTapPlugin", "promptForPushPermission", [showFallbackSettings]);
+}
+
+CleverTap.prototype.isPushPermissionGranted = function(successCallback){
+    cordova.exec(successCallback, null, "CleverTapPlugin", "isPushPermissionGranted", []);
+}
+
 function convertDateToEpochInProperties(items){
 //Conversion of date object in suitable CleverTap format